+++ src/main/groovy/info/solidsoft/gradle/pitest/PitestPluginExtension.groovy
@@ -150,6 +150,7 @@ class PitestPluginExtension {
     final Property<Boolean> enableDefaultIncrementalAnalysis    //specific for Gradle plugin
     final Property<Integer> mutationThreshold
     final Property<Integer> coverageThreshold
+    final Property<Integer> testStrengthThreshold
     final Property<String> mutationEngine
     final Property<Boolean> exportLineCoverage  //for debugging usage only
     final RegularFileProperty jvmPath
@@ -256,6 +257,7 @@ class PitestPluginExtension {
         enableDefaultIncrementalAnalysis = of.property(Boolean)
         mutationThreshold = of.property(Integer)
         coverageThreshold = of.property(Integer)
+        testStrengthThreshold = of.property(Integer)
         mutationEngine = of.property(String)
         exportLineCoverage = of.property(Boolean)
         jvmPath = of.fileProperty()