+++ src/main/groovy/info/solidsoft/gradle/pitest/PitestTask.groovy
@@ -199,6 +199,10 @@ class PitestTask extends JavaExec {
 
     @Input
     @Optional
+    final Property<Integer> testStrengthThreshold
+
+    @Input
+    @Optional
     final Property<String> mutationEngine
 
     @Input
@@ -286,6 +290,7 @@ class PitestTask extends JavaExec {
         defaultFileForHistoryData = of.fileProperty()
         mutationThreshold = of.property(Integer)
         coverageThreshold = of.property(Integer)
+        testStrengthThreshold = of.property(Integer)
         mutationEngine = of.property(String)
         exportLineCoverage = of.property(Boolean)
         jvmPath = of.fileProperty()
@@ -373,6 +378,7 @@ class PitestTask extends JavaExec {
         map['mutableCodePaths'] = (getMutableCodePaths()*.absolutePath)?.join(',')
         map['mutationThreshold'] = optionalPropertyAsString(mutationThreshold)
         map['coverageThreshold'] = optionalPropertyAsString(coverageThreshold)
+        map['testStrengthThreshold'] = optionalPropertyAsString(testStrengthThreshold)
         map['mutationEngine'] = mutationEngine.getOrNull()
         map['exportLineCoverage'] = optionalPropertyAsString(exportLineCoverage)
         map['includeLaunchClasspath'] = Boolean.FALSE.toString()   //code to analyse is passed via classPath