+++ src/test/groovy/info/solidsoft/gradle/pitest/PitestTaskConfigurationSpec.groovy
@@ -51,6 +51,7 @@ class PitestTaskConfigurationSpec extends BasicProjectBuilderSpec implements Wit
                                                                                 'timestampedReports',
                                                                                 'mutationThreshold',
                                                                                 'coverageThreshold',
+                                                                                'testStrengthThreshold',
                                                                                 'mutationEngine',
                                                                                 'exportLineCoverage',
                                                                                 'jvmPath',
@@ -164,6 +165,7 @@ class PitestTaskConfigurationSpec extends BasicProjectBuilderSpec implements Wit
             //enableDefaultIncrementalAnalysis tested separately
             "mutationThreshold"      | 90                                           || "90"
             "coverageThreshold"      | 95                                           || "95"
+            "testStrengthThreshold"  | 95                                           || "95"
             "mutationEngine"         | "gregor2"                                    || "gregor2"
             "exportLineCoverage"     | true                                         || "true"
             "jvmPath"                | new File("//opt//jvm15//")                   || new File("//opt//jvm15//").path